Properly deal with unsupported bootloader
Backend would crash in case it couldn't determine the boot loader type,
which is the case for the live media with syslinux (due to use of
uninitialized value). This adds an assignent of None to it.

In case backend would not crash, frontend would happily attempt to
cofigure the boot loader, despite the fact that there's none (and in
turn, backend would call grubby with incorrect command line). This patch
adds check for it (well, if we don't have a default kernel, we can be
positive that attempt to modify its command line would not succeed).

Furthermore, one unused variable is removed.

Also, this adds a new untranslated string, PO files need to be updated
accordingly.