Allow updating kdump memory without hitting enter
system-config-kdump performs validation of certain input fields and when
this validation succeeds it saves it into its configuration object and
updates other input fields that may depend on it.

The validation is triggered by events such as the input field losing the
input focus or enter key being pressed. For reasons unclear to me, the
events are not run sometimes. For this reason, commit
24d642b4eafde9dc84820177dccd1ccab5956e0b introduced manual invocation of
the validation routines. The kdump memory spin button was not validated
though, and further, the validation routine cannot just be run manually
as it would interfere with the routine for kernel command line.

This change works this around by selectively running the validation
routines for widgets on the current tab when
- the tab is changed,
- apply button/menu entry is pressed.
This way, the widgets on the other tabs are updated based on the current
tab widgets, which were edited the last.

Fixes #877004

Signed-off-by: Martin Milata <mmilata@redhat.com>