Offer partition from kdump.conf as target
It might happen that the partition from kdump.conf is not among those we
extracted from /etc/fstab. This can happen when the partition is
specified by UUID in kdump.conf and by device node in fstab.  Allow the
user to choose this device - if it is unsuitable, the kdump service
restart will fail later.

Possibly resolves: #1146172

Signed-off-by: Martin Milata <mmilata@redhat.com>